moved constants file to cli module and updated imports, delete parameters module
