+++ big_scape/distances/dss.py
@@ -13,7 +13,7 @@ from numpy import ndarray
 from scipy.optimize import linear_sum_assignment
 
 # from other modules
-from big_scape.parameters.constants import LEGACY_ANCHOR_DOMAINS
+from big_scape.cli.constants import LEGACY_ANCHOR_DOMAINS
 from big_scape.comparison import RecordPair
 
 # from this module
