--- big_scape/parameters/__init__.py
@@ -1,22 +1,0 @@
-"""Contains modules for parameters, both user input and constants"""
-from .run import RunParameters
-from .input import InputParameters
-from .output import OutputParameters
-from .comparison import ComparisonParameters
-from .diagnostics import DiagnosticsParameters
-from .hmmer import HmmerParameters
-from .binning import BinningParameters
-from .networking import NetworkingParameters
-from .cmd_parser import parse_cmd
-
-__all__ = [
-    "RunParameters",
-    "InputParameters",
-    "HmmerParameters",
-    "BinningParameters",
-    "ComparisonParameters",
-    "NetworkingParameters",
-    "DiagnosticsParameters",
-    "OutputParameters",
-    "parse_cmd",
-]
